Исправил функции проверки токена и проверки авторизации, все перенесено в роутер, app.js почистил и удалил не нужные фу-ии
